refactor(backend): per-entity route folders, make virtual-api-key canonical
Introduce a documented backend convention — one routes file per entity
under src/routes/<entity>/<entity>.routes.ts, plus per-endpoint test files
named <action>.<entity>.route.test.ts — and convert virtual-api-key to be
its canonical reference. Splits the single 16-test file into per-endpoint
create/get/delete suites and adds backend/CLAUDE.md describing the shape.

Reword platform/CLAUDE.md so its backend guidance points at the new
convention while staying accurate for the entities still on the flat layout.
