+++ platform/backend/src/routes/virtual-api-key/delete.virtual-api-key.route.test.ts
@@ -1,0 +1,93 @@
+import { vi } from "vitest";
+import type { FastifyInstanceWithZod } from "@/server";
+import { createFastifyInstance } from "@/server";
+import { afterEach, beforeEach, describe, expect, test } from "@/test";
+import type { User } from "@/types";
+
+vi.mock("@/auth", () => ({
+  userHasPermission: vi.fn(),
+}));
+
+import { userHasPermission } from "@/auth";
+
+const mockUserHasPermission = vi.mocked(userHasPermission);
+
+describe("DELETE /api/llm-virtual-keys/:id", () => {
+  let app: FastifyInstanceWithZod;
+  let organizationId: string;
+  let user: User;
+
+  beforeEach(async ({ makeOrganization, makeUser }) => {
+    const organization = await makeOrganization();
+    organizationId = organization.id;
+    user = await makeUser();
+    mockUserHasPermission.mockReset();
+    mockUserHasPermission.mockResolvedValue(false);
+
+    app = createFastifyInstance();
+    app.addHook("onRequest", async (request) => {
+      (
+        request as typeof request & {
+          organizationId: string;
+          user: User;
+        }
+      ).organizationId = organizationId;
+      (request as typeof request & { user: User }).user = user;
+    });
+
+    const { default: virtualApiKeysRoutes } = await import(
+      "./virtual-api-key.routes"
+    );
+    await app.register(virtualApiKeysRoutes);
+  });
+
+  afterEach(async () => {
+    await app.close();
+  });
+
+  test("DELETE /api/llm-virtual-keys/:id removes the key", async ({
+    makeLlmProviderApiKey,
+    makeSecret,
+  }) => {
+    mockUserHasPermission.mockResolvedValue(true);
+
+    const secret = await makeSecret({ secret: { apiKey: "sk-real" } });
+    const parentKey = await makeLlmProviderApiKey(organizationId, secret.id, {
+      provider: "openai",
+    });
+
+    const createResponse = await app.inject({
+      method: "POST",
+      url: "/api/llm-virtual-keys",
+      payload: {
+        name: "delete-me",
+        providerApiKeys: [
+          { provider: parentKey.provider, providerApiKeyId: parentKey.id },
+        ],
+      },
+    });
+
+    expect(createResponse.statusCode).toBe(200);
+
+    const deleteResponse = await app.inject({
+      method: "DELETE",
+      url: `/api/llm-virtual-keys/${createResponse.json().id}`,
+    });
+
+    expect(deleteResponse.statusCode).toBe(200);
+    expect(deleteResponse.json()).toEqual({ success: true });
+
+    const listResponse = await app.inject({
+      method: "GET",
+      url: `/api/llm-virtual-keys?providerApiKeyId=${parentKey.id}`,
+    });
+
+    expect(listResponse.statusCode).toBe(200);
+    expect(
+      listResponse
+        .json()
+        .data.map((key: { id: string }) => key.id)
+        .includes(createResponse.json().id),
+    ).toBe(false);
+  });
+});
